blmp: migration strategy: IDLE-PULL
Goals:
1. keep CPUs in faster domains running tasks with a significative
   load ratio

When a CPU on the faster domain goes idle, up-migrate anything which is
above the threshold while running on a CPU of a slower domain.

Looking for a task to be idle-pulled, do not consider tasks where the
affinity does not allow them to be placed on the target CPU, while alsol
ensurig that tasks with restricted affinity do not block selecting other
unrestricted busy tasks.

The "hmp_get_heaviest_task" is used to idenitfy the most effective task
to by idle-pulled from the slower domain into the faster one.

The CPU stopper is used only when the task to be migrated is currently
running the source CPU. Otherwise the task is moved by just locking
source and destination run queues.

This is part of a consolidation of patches originally authored by:
Alex Shi <alex.shi@linaro.org>
Chris Redpath <chris.redpath@arm.com>
Mathieu Poirier <mathieu.poirier@linaro.org>

Signed-off-by: Patrick Bellasi <patrick.bellasi@arm.com>
Signed-off-by: Jon Medhurst <tixy@linaro.org>
Signed-off-by: Jesse Chan <jc@linux.com>
Signed-off-by: Michele Beccalossi <beccalossi.michele@gmail.com>