universal8890: import Samsung OSRC G935SKSU1DQG1
This commit is the kernel source code provided by Samsung to build a
kernel for the Samsung Galaxy S7 and the Samsung Galaxy S7 edge.

This is the result of using the following source code tarball as a base:
 - SM-G935S_KOR_NN_Opensource (G935SKSU1DQG1)
And then merging the missing device configurations from the following
source code tarballs:
 - SM-G930S_KOR_NN_Opensource (G930SKSU1DQG1)
 - SM-G935F_NN_Opensource (G935FXXU1DQG1)

The build output leftovers related to rkp_cfp have been removed.

Signed-off-by: Michele Beccalossi <beccalossi.michele@gmail.com>