ramdisk: replace the fstabs instead of patching them
Due to the size of the patches to fstab.samsungexynos8890 and
fstab.samsungexynos8890.fwup required for f2fs in order to mount, opt to
replace the files with the already patched ones instead.

Currently the changes to the fstabs are equivalent to the ones done by
anykernel.sh before now.

Signed-off-by: Michele Beccalossi <beccalossi.michele@gmail.com>